move api.tests.clouds.aws.test_models.ModelStrTestMixin to api.tests.helper.ModelStrTestMixin for cleaner reusability
